image controller to image model
